requirements.txt
azure-storage-blob==12.16.0
beautifulsoup4==4.12.0
pandas==2.0.3
psycopg2==2.9.6
pyspark==3.4.1
Requests==2.31.0
